# loyalty-engine/openapi.yaml
openapi: 3.0.0
info:
  title: Loyalty Engine API
  version: 1.0.0
  description: An API for retrieving user rewards and details.

paths:
  /rewards:
    get:
      summary: Get Reward Offers
      responses:
        '200':
          description: Successful response with reward offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RewardOffer'
  /user-rewards:
    get:
      summary: Get all user rewards.
      responses:
        '200':
          description: A list of user rewards.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserReward'
  /rewards/{id}:
    get:
      summary: Get user details by their ID.
      parameters:
      - name: id
        in: path
        description: Reward ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: User details for the given ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardOffer'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardOffer'

  /user/{id}:
    get:
      summary: Get user details by their ID.
      parameters:
      - name: id
        in: path
        description: User ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: User details for the given ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          
  /reward-confirmation:
    get:
      summary: Get reward confirmation.
      parameters:
      - name: userId
        in: query
        description: User ID
        required: true
        schema:
          type: string
      - name: rewardId
        in: query
        description: Reward ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Reward confirmation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardConfirmation'
        '500':
          description: Internal server error.
          content:
            text/plain:
              schema:
                type: string
                example: "failed to fetch reward confirmation"
  /reward-confirmations:
    get:
      summary: Get reward confirmations.
      parameters:
      - name: userId
        in: query
        description: User ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: A list of reward confirmations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RewardConfirmationWithQR'
        '500':
          description: Internal server error.
          content:
            text/plain:
              schema:
                type: string
                example: "failed to fetch reward confirmation"
  /qr-code:
    get:
      summary: Get QR code.
      parameters:
      - name: userId
        in: query
        description: User ID
        required: true
        schema:
          type: string
      - name: rewardId
        in: query
        description: Reward ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: QR code.
          content:
            image/png:
              schema:
                type: object
        '500':
          description: Internal server error.
          content:
            text/plain:
              schema:
                type: string
                example: "failed to fetch QR code"
  /qr-codes:
    get:
      summary: Get QR codes.
      parameters:
      - name: userId
        in: query
        description: User ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Status OK.
        '500':
          description: Internal server error.
          content:
            text/plain:
              schema:
                type: string
                example: "failed to fetch QR code"
components:
  schemas:
    User:
      type: object
      properties:
        userId:
          type: string
          description: Unique user identifier.
        firstName:
          type: string
          description: User's first name.
        lastName:
          type: string
          description: User's last name.
        email:
          type: string
          description: User's email address.
          format: email
    RewardOffer:
      type: object
      properties:
        Id:
          type: string
          description: Reward offer ID
        Name:
          type: string
          description: Reward offer name
        Value:
          type: number
          format: float
          description: Reward offer value
        TotalPoints:
          type: integer
          description: Total points required for the reward
        Description:
          type: string
          description: Reward offer description
        LogoUrl:
          type: string
          description: Reward offer logo URL
    UserReward:
      type: object
      properties:
        userId:
          type: string
          description: Unique user identifier.
        selectedRewardDealId:
          type: string
          description: ID of the selected reward deal.
        timestamp:
          type: string
          description: Timestamp of the reward selection.
          format: date-time
        acceptedTnC:
          type: boolean
          description: Indicates if user accepted the Terms and Conditions.
    RewardConfirmation:
      type: object
      properties:
        userId:
          type: string
          description: Unique user identifier.
        rewardId:
          type: string
          description: ID of the reward.
        rewardConfirmationNumber:
          type: string
          description: Reward confirmation number.
    RewardConfirmationWithQR:
      type: object
      properties:
        userId:
          type: string
          description: Unique user identifier.
        rewardId:
          type: string
          description: ID of the reward.
        qrCode:
          type: string
          description: Encoded QR code.
    