# reward-management-web-app/src/api/types.ts
export interface CardDetails{
    userId: string;
    cardNumber: string;
    rewardPoints: number
    currentBalance: number
    dueAmount: number
    lastStatementBalance: number
    availableCredit: number
}

export interface Reward{
    id: string;
    name: string;
    logoUrl: string;
    value: number;
    totalPoints: number;
    description?: string;
}

export interface RewardConfirmation{
    userId: string;
    rewardId: string;
    qrCode: string;
}